fix UI language switcher, and seperate UI and logic for firstForm
