+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Layout from "./components/common/Layout";
+import Layout from "./components/common/Layout/Layout";
 
 function App() {
   return <Layout />;
