+++ src/components/common/Layout/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet, useLocation } from "react-router-dom";
-import Header from "./Header/Header";
-import Footer from "./Footer";
+import Header from "../Header/Header";
+import Footer from "../Footer/Footer";
 
 const Layout = () => {
   const location = useLocation();
