+++ src/components/pages/FirstFormPage/FirstFormPage.tsx
@@ -1,5 +1,5 @@
 import FirstForm from "./FirstForm";
-import usePageTitle from "../../../hooks/usePageTitle";
+import usePageTitle from "../../../hooks/pageTilte/usePageTitle";
 import { useTranslation } from "react-i18next";
 
 export default function FirstFormPage() {
