+++ src/components/pages/NinthFormPage/NinthFormPage.tsx
@@ -1,4 +1,4 @@
-import usePageTitle from "../../../hooks/usePageTitle";
+import usePageTitle from "../../../hooks/pageTilte/usePageTitle";
 import { useNavigate } from "react-router-dom";
 import {
   Box,
