+++ src/components/pages/SecondFormPage/SecondFormPage.tsx
@@ -4,7 +4,7 @@ import { Box, IconButton } from "@mui/material";
 import ArrowBackIosNewIcon from "@mui/icons-material/ArrowBackIosNew";
 import ArrowForwardIosIcon from "@mui/icons-material/ArrowForwardIos";
 import SecondForm from "./SecondForm";
-import usePageTitle from "../../../hooks/usePageTitle";
+import usePageTitle from "../../../hooks/pageTilte/usePageTitle";
 import { useTranslation } from "react-i18next";
 import { useFormData } from "../../../contexts/FormData/useFormData";
