+++ src/components/pages/SeventhFormPage/SeventhFormPage.tsx
@@ -1,4 +1,4 @@
-import usePageTitle from "../../../hooks/usePageTitle";
+import usePageTitle from "../../../hooks/pageTilte/usePageTitle";
 import { useTranslation } from "react-i18next";
 import SeventhForm from "./SeventhForm";
 import { Box, IconButton } from "@mui/material";
