+++ src/components/pages/ThirdFormPage/ThirdFormPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import ThirdFormMain from "./ThirdFormMain";
-import usePageTitle from "../../../hooks/usePageTitle";
+import usePageTitle from "../../../hooks/pageTilte/usePageTitle";
 import { useTranslation } from "react-i18next";
 import ThirdFormOptional from "./ThirdFormOptional";
 import { Box, IconButton } from "@mui/material";
